% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abortion.R
\name{abortion}
\alias{abortion}
\title{Abortion}
\usage{
abortion(
  result,
  residual = NULL,
  events = NULL,
  abortion.chance = 0.0029,
  biweekly = FALSE
)
}
\arguments{
\item{result}{The result from the previous day or modified result
from the current day.}

\item{residual}{The rows with timer == 0 that still need to be
processed. The default is NULL if there is nothing to be done}

\item{events}{The events that have already been generated. The
default is NULL in the case that no events have yet been
generated for this day}

\item{abortion.chance}{The probability of abortion per day}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
A list of the (modified) result, the new residual and the
    new events appended to the events you fed to the function
}
\description{
In this function we move bred animals from the gestation section
to a breeding buffer section or cull them.
}
