% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{as_u_matrix}
\alias{as_u_matrix}
\title{as_u_matrix}
\usage{
as_u_matrix(x)
}
\arguments{
\item{x}{A trajectory}
}
\value{
A u matrix
}
\description{
Coerce the data.frame to a transposed u matrix.
}
