% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{as_v_matrix}
\alias{as_v_matrix}
\title{as_v_matrix}
\usage{
as_v_matrix(x)
}
\arguments{
\item{x}{a trajectory}
}
\value{
a vmatrix
}
\description{
Coerce the data.frame to a transposed v matrix.
}
