% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning.R
\name{check.pens}
\alias{check.pens}
\title{check.pens}
\usage{
check.pens(animals, dest.pen)
}
\arguments{
\item{animals}{integer of the amount of animals in departure pen}

\item{dest.pen}{vector of pen ids}
}
\value{
boolean
}
\description{
function to check if there is space in farrowing pens
}
