% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cf.R
\name{clean_cf}
\alias{clean_cf}
\title{clean_cf}
\usage{
clean_cf(result, events, time = 1)
}
\arguments{
\item{result}{dataframe of the herd structure (from one day)}

\item{events}{dataframe of pre-generated base events}

\item{time}{integer, time when we want the cleaning events to start}
}
\value{
dataframe, merged from base events and added cleaning events
}
\description{
Generates the events for cleaning for continuous flow pens
based on weaning (happens currectly every 7 days)
}
