% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{clean_trajectory}
\alias{clean_trajectory}
\title{clean trajectory}
\usage{
clean_trajectory(x)
}
\arguments{
\item{x}{the object$result of a trajectory}
}
\value{
dataframe
}
\description{
format the daily trajectory into more readable
}
