% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{cleaning}
\alias{cleaning}
\title{Cleaning}
\usage{
cleaning(model, result, time, inf_time, events, clean_AIAO, clean_CF)
}
\arguments{
\item{model}{model object}

\item{result}{dataframe. "non-cleaned" trajectory of the model}

\item{time}{A time series}

\item{inf_time}{Time when gilts were infected}

\item{events}{pregenerated events with infected animals}

\item{clean_AIAO}{alternative infectious pressure after cleaning,  all-in-all-out pens}

\item{clean_CF}{alternative infectious pressure after cleaning, continuous flow pens}
}
\value{
model object
}
\description{
run the model with premade events with different cleaning regimes
}
