% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u0.R
\name{create_u0}
\alias{create_u0}
\title{Generates the herd structure}
\usage{
create_u0(
  sow.breeding.pens.persection = 10,
  sow.breeding.pens.section = 1,
  gilt.breeding.pens.persection = 8,
  gilt.breeding.pens.section = 1,
  sow.breeding.buffer.pens.persection = 5,
  sow.breeding.buffer.pens.section = 1,
  gilt.breeding.buffer.pens.persection = 10,
  gilt.breeding.buffer.pens.section = 1,
  sow.gestation.pens.persection = 35,
  sow.gestation.pens.section = 1,
  gilt.gestation.pens.persection = 30,
  gilt.gestation.pens.section = 1,
  farrowing.pens.persection = 26,
  farrowing.pens.sections = 6,
  growing.pens.persection = 26,
  growing.pens.sections = 10,
  growing.buffer.pens.persection = 26,
  growing.buffer.pens.sections = 1,
  finishing.pens.persection = 30,
  finishing.pens.sections = 18,
  growing.gilt.pens.persection = 25,
  growing.gilt.pens.sections = 1,
  node = FALSE,
  time = FALSE,
  capacity = FALSE,
  npigs = FALSE,
  phi = FALSE,
  s_phi = FALSE,
  f_phi = FALSE,
  test_indicator = FALSE,
  result_indicator = FALSE,
  cleaning_indicator = FALSE,
  Icum = FALSE
)
}
\arguments{
\item{sow.breeding.pens.persection}{integer number of pens in the section}

\item{sow.breeding.pens.section}{integer number of sections in the unit}

\item{gilt.breeding.pens.persection}{integer number of pens in the section}

\item{gilt.breeding.pens.section}{integer number of sections in the unit}

\item{sow.breeding.buffer.pens.persection}{integer number of pens in the section}

\item{sow.breeding.buffer.pens.section}{integer number of sections in the unit}

\item{gilt.breeding.buffer.pens.persection}{integer number of pens in the section}

\item{gilt.breeding.buffer.pens.section}{integer number of sections in the unit}

\item{sow.gestation.pens.persection}{integer number of pens in the section}

\item{sow.gestation.pens.section}{integer number of sections in the unit}

\item{gilt.gestation.pens.persection}{integer number of pens in the section}

\item{gilt.gestation.pens.section}{integer number of sections in the unitn}

\item{farrowing.pens.persection}{integer number of pens in the section}

\item{farrowing.pens.sections}{integer number of sections in the unit}

\item{growing.pens.persection}{integer number of pens in the section}

\item{growing.pens.sections}{integer number of sections in the unit}

\item{growing.buffer.pens.persection}{integer number of pens in the section}

\item{growing.buffer.pens.sections}{integer number of sections in the unit}

\item{finishing.pens.persection}{integer number of pens in the section}

\item{finishing.pens.sections}{integer number of sections in the unit}

\item{growing.gilt.pens.persection}{integer number of pens in the section}

\item{growing.gilt.pens.sections}{integer number of sections in the unit}

\item{node}{include node id?}

\item{time}{include time = 1?}

\item{capacity}{include pen capacity?}

\item{npigs}{include npigs?}

\item{phi}{include phi?}

\item{s_phi}{include s_phi?}

\item{f_phi}{include f_phi?}

\item{test_indicator}{include test_indicator?}

\item{result_indicator}{include test_indicator?}

\item{cleaning_indicator}{include test_indicator?}

\item{Icum}{include test_indicator?}
}
\value{
dataframe of herd structure
}
\description{
Generates the herd structure
}
