% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{daily}
\usage{
daily(
  result,
  model,
  tspan,
  crossfostering_day = 34,
  crossfostering_proportion = 0.1,
  finisher_mixing_proportion = 1,
  biweekly = FALSE
)
}
\arguments{
\item{result}{The result object}

\item{model}{The model object}

\item{tspan}{The time vector}

\item{crossfostering_day}{The day to perform crossfostering. The
default is day 34. This both prevents mortality on that day
and does the crossfostering on that day.}

\item{crossfostering_proportion}{The proportion of piglets to be crossfostered}

\item{finisher_mixing_proportion}{The proportion of finisher pigs to be mixed}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
An object for the day
}
\description{
This function runs once per day to generate the day's events
}
