% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{disease_testing}
\alias{disease_testing}
\title{Disease testing}
\usage{
disease_testing(model, events, time, test_weaning, test_gilts)
}
\arguments{
\item{model}{model object}

\item{events}{dataframe of base events with infected animals}

\item{time}{timepoint (single value) when we want testing events to start}

\item{test_weaning}{if TRUE, test sows at the time of weaning}

\item{test_gilts}{if TRUE, test new gilts in the gilt unit before moving them to breeding}
}
\value{
list of result dataframes for different testing scenarios
}
\description{
Disease testing
}
