% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_herd.R
\name{empty.pens}
\alias{empty.pens}
\title{empty.pens}
\usage{
empty.pens(
  x,
  pentype = c("Sow breeding", "Gilt breeding", "Sow breeding buffer",
    "Gilt breeding buffer", "Sow gestation", "Gilt gestation", "Farrowing", "Growing",
    "Growing buffer", "Finishing", "Gilt growing"),
  events = NULL,
  countdown_check = TRUE
)
}
\arguments{
\item{x}{A dataframe which is the output from the trajectory of
the MRSA model from 1 day.}

\item{pentype}{A string}

\item{events}{The queued events for this timestep}

\item{countdown_check}{If TRUE check the countdown timer is 99999}
}
\value{
a vector of pen ids
}
\description{
A function that identifies empty pens of a given class.  main
functionality has been moved to the sectioning-functions below,
which has the same functionality but takes the sections into
account.  Currently used for moving animals from farrowing to
breeding because we group up animals from one pen to one pen.
}
