% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{event}
\alias{event}
\title{event}
\usage{
event(type, time, node, dest, n, proportion)
}
\arguments{
\item{type}{A value that can be coersed to an integer between 1
    and 20. The following values are available:

1 Move a sow from breeding to gestation/the opposite direction: select = 1 shift = 0
2 Move a sow from gestation to farrowing: select = 1 shift = 0
3 Move a sow from farrowing to breeding: select = 1 shift = 0
4 Move a gilt grower to breeding: select = 2 shift = 0
5 Move a gilt from breeding to gestation/the opposite direction: select = 2 shift = 0
6 Move a gilt from gestation to farrowing: select = 2 shift = 1
7 Piglets born: select = 3 shift = 0
8 Piglets weaned to grower: select = 4 shift = 2
9 Growers moved to finisher: select = 5 shift = 3
10 Finishers to slaughter: select= 6 shift = 0
11 piglets move to gilt grower (-> gilts): select = 4, shift = 4
12 piglet mortality: select = 4, shift = 0
13 grower mortality: select = 5, shift = 0
14 sow mortality: select = 1, shift = 0
15 gilt mortality: select = 2, shift = 0
16 Purchase gilt: select = 7, shift = 0
17 Gilt growers to finishing (intTrans): select = 5, shift = 3 ##NOT IN USE
18 Growers moved to growing buffer: select = 5, shift = 0
19 Finishers moved to finishing buffer: select = 6, shift = 0
20 Purchase infected gilt: select = 8, shift = 0
21 Mingle piglets: select = 4, shift = 0
22 Mingle finishers: select = 6 shift = 0
23 Event that infects pigs: select = 9 shift = 5
24 test indicator: select = 10, shift = 0
25 both indicators: select = 11, shift = 0}

\item{time}{The time}

\item{node}{The node}

\item{dest}{The destination node}

\item{n}{The number of animals}

\item{proportion}{The proportion of animals}
}
\value{
A data.frame with one line
}
\description{
A function that give back an event but doesn't require us to
understand what shift and select should be.
}
