% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{event_dataframe}
\alias{event_dataframe}
\title{event_dataframe}
\usage{
event_dataframe(
  event = c("extTrans", "intTrans", "enter", "exit"),
  time = 1,
  node,
  dest,
  n,
  proportion = 0,
  select = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11),
  shift = c(1, 2, 3, 4, 5)
)
}
\arguments{
\item{event}{The event type; one of "extTrans", "intTrans",
"enter", "exit".}

\item{time}{A value that can be coersed to an integer greater than
or equal to 1.}

\item{node}{A value that can be coersed to an integer greater than
or equal to 1.}

\item{dest}{A value that can be coersed to an integer greater than
or equal to 0.}

\item{n}{A value that can be coersed to an integer greater than or
equal to 0.}

\item{proportion}{A numeric between 0 and 1}

\item{select}{The select value; an value that can be coersed to an
integer between 1 and 6}

\item{shift}{The shift value; an value that can be coersed to an
integer between 0 and 4}
}
\value{
A dataframe
}
\description{
A function to generate a single event. This is a dataframe with
one line in the correct structure that we expect for events.
}
