% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Events for baseline model}
\format{
## `events`
\describe{
  \item{event}{Event type}
  \item{time}{Time of when the event occurs}
  \item{node}{The node that the event operates on}
  \item{dest}{The destination node}
  \item{n}{The number of individuals}
  \item{select}{Column from E matrix}
  \item{shift}{Column from N matrix}
}
}
\usage{
events
}
\description{
Events for baseline model
}
\keyword{datasets}
