% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events_cf34_cfp10_fmp0}
\alias{events_cf34_cfp10_fmp0}
\title{Model events with 10\% cross-fostering and 0\% finishing mixing}
\format{
## `events_cf34_cfp10_fmp0`
\describe{
  \item{event}{Event type}
  \item{time}{Time of when the event occurs}
  \item{node}{The node that the event operates on}
  \item{dest}{The destination node}
  \item{n}{The number of individuals}
  \item{select}{Column from E matrix}
  \item{shift}{Column from N matrix}
}
}
\usage{
events_cf34_cfp10_fmp0
}
\description{
Model events with 10\% cross-fostering and 0\% finishing mixing
}
\keyword{datasets}
