% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events_cf34_cfp10_fmp100_biweekly}
\alias{events_cf34_cfp10_fmp100_biweekly}
\title{Biweekly model with baseline mixing}
\format{
## `events_cf34_cfp10_fmp100_biweekly`
\describe{
  \item{event}{Event type}
  \item{time}{Time of when the event occurs}
  \item{node}{The node that the event operates on}
  \item{dest}{The destination node}
  \item{n}{The number of individuals}
  \item{select}{Column from E matrix}
  \item{shift}{Column from N matrix}
}
}
\usage{
events_cf34_cfp10_fmp100_biweekly
}
\description{
Biweekly model with baseline mixing
}
\keyword{datasets}
