% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events_mixing_cf0_fm0_d34_at1500}
\alias{events_mixing_cf0_fm0_d34_at1500}
\title{Model events introduced at day 1500 with 0\% cross-fostering and 0\% finishing mixing}
\format{
## `events_mixing_cf0_fm0_d34_at1500`
\describe{
  \item{event}{Event type}
  \item{time}{Time of when the event occurs}
  \item{node}{The node that the event operates on}
  \item{dest}{The destination node}
  \item{n}{The number of individuals}
  \item{select}{Column from E matrix}
  \item{shift}{Column from N matrix}
}
}
\usage{
events_mixing_cf0_fm0_d34_at1500
}
\description{
Model events introduced at day 1500 with 0\% cross-fostering and 0\% finishing mixing
}
\keyword{datasets}
