% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_image.R
\name{farm_polygons}
\alias{farm_polygons}
\title{farm_polygons}
\usage{
farm_polygons(result = NULL)
}
\arguments{
\item{result}{a result of a trajectory}
}
\value{
A polygon object
}
\description{
Create the farm polygons
}
