% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing_slaughter.R
\name{finishing_slaughter}
\alias{finishing_slaughter}
\title{finishing_slaughter}
\usage{
finishing_slaughter(result, residual = NULL, events = NULL, biweekly = FALSE)
}
\arguments{
\item{result}{The result object}

\item{residual}{The remaining rows to process for the day}

\item{events}{Potential events to pas through}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
An object with the result dataframe, events and residual
}
\description{
finishing_slaughter
}
