% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit}
\alias{fit}
\title{Model fit obtained from Approximate Bayesian Computation which was used for parameterising transmission rates from published data}
\format{
## `fit`
\describe{
The fit includes for example the model object (model), the accepted transmission rates per generation (x),
as well metrics about the parameterisation process such as number of proposed particles per generations (nprop).
}
}
\usage{
fit
}
\description{
Model fit obtained from Approximate Bayesian Computation which was used for parameterising transmission rates from published data
}
\keyword{datasets}
