% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tests.R
\name{gen_tests}
\alias{gen_tests}
\title{gen_tests}
\usage{
gen_tests(events, weaning = TRUE, gilts = TRUE, time = 1)
}
\arguments{
\item{events}{dataframe, consists of generated movement events}

\item{weaning}{boolean, generate testing events for sows at weaning}

\item{gilts}{boolean, generate testing events for growing gilts before moving to breeding}

\item{time}{integer, timepoint when we want the testing events to start}
}
\value{
An events dataframe where testing events are merged with old events.
}
\description{
Generate events for disease testing of sows or growing gilts.
Sows are tested before they are moved from farrowing to breeding.
Growing gilts are tested before they are moved from gilt growing
unit to gilt breeding.
}
