% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilts_to_slaughter.R
\name{gilts_to_slaughter}
\alias{gilts_to_slaughter}
\title{gilts_to_slaughter}
\usage{
gilts_to_slaughter(result, residual = NULL, events = NULL, biweekly = FALSE)
}
\arguments{
\item{result}{The result of a trajectory}

\item{residual}{the remaining nodes to process}

\item{events}{an event dataframe}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
An object to pass to the next event generator
}
\description{
gilts_to_slaughter
}
