% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_image.R
\name{grid_to_polygons}
\alias{grid_to_polygons}
\title{grid_to_polygons}
\usage{
grid_to_polygons(grid)
}
\arguments{
\item{grid}{A grid object}
}
\value{
A grid of polygons
}
\description{
Convert a grid to polygons
}
