% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growing_buffer_to_finishing.R
\name{growing_buffer_to_finishing}
\alias{growing_buffer_to_finishing}
\title{Growing buffer to finishing}
\usage{
growing_buffer_to_finishing(
  result,
  residual = NULL,
  events = NULL,
  biweekly = FALSE
)
}
\arguments{
\item{result}{The result from the previous day or modified result
from the current day.}

\item{residual}{The rows with timer == 0 that still need to be
processed. The default is NULL if there is nothing to be done}

\item{events}{The events that have already been generated. The
default is  NULL in the case that no events have yet been
generated for this day}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
list of the (modified) result, the new residual and the
    new events appended to the events you fed to the function
}
\description{
In this function we merge the pigs in buffer pens. It prioritizes merging with
a finishing section that already has pigs and the section that has the youngest
finishing pigs.
}
