% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect_pigs.R
\name{infect_pigs}
\alias{infect_pigs}
\title{infect_pigs}
\usage{
infect_pigs(x, time, pentype = "Gilt growing", proportion = 0.2)
}
\arguments{
\item{x}{result dataframe from the herd}

\item{time}{time when events are scheduled}

\item{pentype}{A string of the pentype. This could be any valid
pen type the default is "Gilt growing"}

\item{proportion}{proportion of animals in each node to be
infected}
}
\value{
dataframe of scheduled events
}
\description{
A function to infect pigs using events
}
