% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_herd.R
\name{initialize_herd}
\alias{initialize_herd}
\title{initialize_herd}
\usage{
initialize_herd(
  result,
  residual = NULL,
  events = NULL,
  biweekly = FALSE,
  gilt_introduction = 21
)
}
\arguments{
\item{result}{The result from the previous day or modified result
from the current day.}

\item{residual}{The rows with timer == 0 that still need to be
processed. The default is NULL if there is nothing to be done}

\item{events}{The events that have already been generated. The
default is NULL in the case that no events have yet been
generated for this day}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}

\item{gilt_introduction}{Change how many times gilts are introduced
(based on gilt_interval). Affects the end herd size.}
}
\value{
A list of the (modified) result, the new residual and the
    new events appended to the events you fed to the function
}
\description{
In this function we add gilts to the breeding at weekly intervals
in the beginning of the study to startup to herd.
}
