% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{interventions}
\alias{interventions}
\title{Interventions}
\usage{
interventions(
  intervention = "basemodel",
  events,
  proportion = 0.2,
  time = 731:3000,
  clean_AIAO = 1,
  clean_CF = 1,
  test_weaning = FALSE,
  test_gilts = FALSE,
  Se = 0.7,
  room_scale = 0.1,
  farm_scale = 0.1
)
}
\arguments{
\item{intervention}{string, which intervention model to prepare}

\item{events}{pre-generated events}

\item{proportion}{proportion of gilts to be infected}

\item{time}{timeseries for model run}

\item{clean_AIAO}{alternative infectious pressure after cleaning, all-in-all-out pens}

\item{clean_CF}{alternative infectious pressure after cleaning, continuous flow pens}

\item{test_weaning}{if TRUE, test sows at the time of weaning}

\item{test_gilts}{if TRUE, test new gilts in the gilt unit before moving them to breeding}

\item{Se}{The sensitivity of the test applied in the sows and gilts}

\item{room_scale}{between_pen spread within a room}

\item{farm_scale}{between room spread within the herd}
}
\value{
a model object
}
\description{
Interventions
}
