% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrowing.R
\name{littersize}
\alias{littersize}
\title{littersize}
\usage{
littersize(litters, size = 14.8)
}
\arguments{
\item{litters}{an integer for the amount of litters needed}

\item{size}{an integer lambda for poisson distribution (size of litter)}
}
\value{
a vector of integers
}
\description{
sampling the littersizes with poisson distribution
poisson is used instead of binomial distribution to better reflect the chance of getting littersizes outsife of SD
}
