% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_image.R
\name{make_grid}
\alias{make_grid}
\title{make_grid}
\usage{
make_grid(width, height, scale.x = 1, scale.y = 1, reference_pt = c(0, 0))
}
\arguments{
\item{width}{The total width in number of grid subunits}

\item{height}{The total height in number of grid subunits}

\item{scale.x}{The size of the x units of the grid}

\item{scale.y}{The size of the y units of the grid}

\item{reference_pt}{The bottom left corner of the grid. A vector
of length 2 with the x and y coordinate.}
}
\value{
a matrix with points on a grid
}
\description{
a utility function to make a grid
}
