% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept.R
\name{MRSA_distance}
\alias{MRSA_distance}
\title{MRSA_distance}
\usage{
MRSA_distance(result, expected, ...)
}
\arguments{
\item{result}{The ruesult of a run trajectory of the model}

\item{expected}{The expected results}

\item{...}{Other arguments}
}
\value{
A number
}
\description{
calculate the distance between the trajectory and the expected
data.
}
