% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept.R
\name{MRSA_expected}
\alias{MRSA_expected}
\title{MRSA_expected}
\usage{
MRSA_expected()
}
\value{
A dataframe containing the expected values to parameterize
    the model.
}
\description{
The expected prevalence at various time points used to
parameterize the model. There are 8 target observation points in
the herd that are included in the data. These are:
}
\details{
* q_m1s : is the expected prevalence in sows when they enter the
    farrowing unit.

* q_m2s : the expected prevalence in the sows three days after
    entering the farrowing unit.

* q_m2p : the expected prevalence in the piglets three days after
    entering the farrowing unit.

* q_m3s : the expected prevalence in the sows three weeks after
    entering the farrowing unit.

* q_m3p : the expected prevalence in the piglets three weeks after
    entering the farrowing unit.

* q_m4p : the expected prevalance in the pigs 7 days after weaning

* q_m5p : the expected prevalance in the pigs 35 days after weaning

* q_m6p : the expected prevalence in the pigs 84 days after moving
    the the finishing unit

For each of the observation there was variation in the data from
the Broens paper. We therefore supply the 10th 50th and 90th
percentiles of those observations to mimic what you might see in a
low prevalence, 'normal' and high prevalence herd repectively.
}
