% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{MRSA_model}
\alias{MRSA_model}
\title{MRSA_model}
\usage{
MRSA_model(
  Nt = 0.8705506,
  beta_phi = 0.00285,
  duration = 17.4,
  mortality_grower = 4e-04,
  mortality_finish = 2e-04,
  room_scale = 0.1,
  farm_scale = 0.1,
  cleaning_effect_AIAO = 1,
  cleaning_effect_CF = 1,
  Se = 0.7
)
}
\arguments{
\item{Nt}{The daily decay rate (The proporation of phi that is
left after one day's decay)}

\item{beta_phi}{The indirect transmission rate for all animal categories}

\item{duration}{The mean duration of carriage. This is estimated
as a erlang distribution with shape parameter = 3.}

\item{mortality_grower}{The daily mortality rate in growers}

\item{mortality_finish}{The daily mortality rate in finishers}

\item{room_scale}{The relative effect of phi at the room level compared to phi at the pen level.}

\item{farm_scale}{The relative effect of phi at the farm level compared to phi at the room level.}

\item{cleaning_effect_AIAO}{The proportion of phi left after cleaning AIAO pens}

\item{cleaning_effect_CF}{The proportion of phi left after cleaning any pens}

\item{Se}{The diagnostic sensitivity of the test}
}
\value{
A SimInf model object
}
\description{
MRSA_model
}
