% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_4_parameter.R
\name{MRSA_model_4_parameter}
\alias{MRSA_model_4_parameter}
\title{MRSA_model_4_parameter}
\usage{
MRSA_model_4_parameter(
  Nt = 0.8705506,
  beta_phi = 0.00285,
  beta_phi_mature = 8e-04,
  beta_phi_piglets = 8e-04,
  beta_phi_growers = 8e-04,
  beta_phi_finish = 8e-04,
  duration = 17.4,
  mortality_grower = 4e-04,
  mortality_finish = 2e-04,
  room_scale = 0.1,
  farm_scale = 0.1,
  cleaning_effect_AIAO = 1,
  cleaning_effect_CF = 1,
  Se = 0.7
)
}
\arguments{
\item{Nt}{The daily decay rate (The proporation of phi that is
left after one day's decay)}

\item{beta_phi}{Not used}

\item{beta_phi_mature}{The indirect transmission rate for sows and gilts}

\item{beta_phi_piglets}{The indirect transmission rate for piglets}

\item{beta_phi_growers}{The indirect transmission rate for growers}

\item{beta_phi_finish}{The indirect transmission rate for finishers}

\item{duration}{The mean duration of carriage. This is estimated
as a erlang distribution with shape parameter = 3.}

\item{mortality_grower}{The daily mortality rate in growers}

\item{mortality_finish}{The daily mortality rate in finishers}

\item{room_scale}{The relative effect of phi at the room level compared to phi at the pen level.}

\item{farm_scale}{The relative effect of phi at the farm level compared to phi at the room level.}

\item{cleaning_effect_AIAO}{The proportion of phi left after cleaning AIAO pens}

\item{cleaning_effect_CF}{The proportion of phi left after cleaning any pens}

\item{Se}{The diagnostic sensitivity of the test}
}
\value{
A SimInf model object
}
\description{
MRSA_model_4_parameter
}
