% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{pen_capacity}
\alias{pen_capacity}
\title{pen_capacity}
\usage{
pen_capacity(pentype)
}
\arguments{
\item{pentype}{A character vector of pentypes}
}
\value{
an numeric vector of the capacities of these pentypes
}
\description{
pensize of a pentype
}
