% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_image.R
\name{pen_polygon}
\alias{pen_polygon}
\title{pen_polygon}
\usage{
pen_polygon(coords, ID = "poly1")
}
\arguments{
\item{coords}{a matrix of dim c(4, 2) containing the x and y
coordinates of the 4 corners of the polygon in
counterclockwise order}

\item{ID}{The id of the poly}
}
\value{
A polygon object
}
\description{
A function that create a polygon object from coordinates
}
