% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_image.R
\name{pen_polygons}
\alias{pen_polygons}
\title{pen_polygons}
\usage{
pen_polygons(
  width,
  height,
  scale.x = 1,
  scale.y = 1,
  reference_pt = c(0, 0),
  IDS = NULL
)
}
\arguments{
\item{width}{number of pens wide}

\item{height}{number of pens high}

\item{scale.x}{The width of each pen}

\item{scale.y}{The height of each pen}

\item{reference_pt}{The coordinate of the bottom left of the group
of pens}

\item{IDS}{The ids of the pens}
}
\value{
A group of pens
}
\description{
A function to create a group of pens:
}
