% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piglet_mortality.R
\name{piglet_mortality}
\alias{piglet_mortality}
\title{piglet_mortality}
\usage{
piglet_mortality(
  result,
  residual = NULL,
  events = NULL,
  prob = 0.006,
  skip_days = 34
)
}
\arguments{
\item{result}{A result from a model trajectory}

\item{residual}{The residual}

\item{events}{The events}

\item{prob}{The fraction of piglet to die}

\item{skip_days}{A vector of days to skip for mortality. This is
used to prevent mortality on the same days as mixing since it
interferes.}
}
\value{
An events dataframe
}
\description{
function to handle the piglet mortality outside siminf prob is the
probability of death for each piglet per day, 0.006 when piglets
spend 33 d in the unit (total 35 d - 1 day for weaning - 1 day to
ignore the first day of life because of piglets mixing) total
mortality between birth and weaning according to 2019 statistics is
17.7
}
