% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy.R
\name{pregnancy}
\alias{pregnancy}
\title{pregnancy}
\usage{
pregnancy(pigs, failure_rate = 0.055)
}
\arguments{
\item{pigs}{an integer of the amount of animals in pen}

\item{failure_rate}{The proportion of failures}
}
\value{
an integer representing the amount of animals that are not
    pregnant in the given pen
}
\description{
Function to sample how many animals in given pen are not
pregnant. Sums the amount of animals of given category from given
node. Mean omloepningsprocent is 5.5 % in Winpig at 2019
}
