% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{scale_p}
\alias{scale_p}
\title{scale p}
\usage{
scale_p(dest, p)
}
\arguments{
\item{dest}{The ID(s) of destination node(s)}

\item{p}{A vector of the proportion to put in each destination
node}
}
\value{
A vector of proportions to be passed to the events
}
\description{
Scaling proportions when we are sampling for production phase that
works with proportions instead of n
}
