% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectioning.R
\name{sectioning}
\alias{sectioning}
\title{sectioning}
\usage{
sectioning(
  x,
  pentype = c("Sow breeding", "Gilt breeding", "Sow breeding buffer",
    "Gilt breeding buffer", "Sow gestation", "Gilt gestation", "Farrowing", "Growing",
    "Growing buffer", "Finishing", "Finishing buffer", "Gilt growing"),
  pen.total = 0,
  events = NULL
)
}
\arguments{
\item{x}{A dataframe which is the output from the trajectory of
the MRSA model from 1 day.}

\item{pentype}{A string}

\item{pen.total}{integer for required amount of pens, if it's 0,
we want to have a whole empty section}

\item{events}{The queued events for this timestep}
}
\value{
vector of pen ids in the given pentype that have enough
    empty pens
}
\description{
In this function animals are moved one by one (n =1) from one pen
into several pens allows specifying if we want the destination
section to be fully empty (all-in-all-out) or just have x number
of pens used in: sows/gilts from gestation to farrowing,
sows/gilts from breeding to gestation This function checks also if
there are already events scheduled for the destination pens and
removes these from the returned pen list.  Returns a vector of pen
ids in given pentype that have enough empty pens in one section.
}
