% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_select.R
\name{select_matrix}
\alias{select_matrix}
\title{select_matrix}
\usage{
select_matrix(compartments = NULL)
}
\arguments{
\item{compartments}{A vector of compartment}
}
\value{
a matrix
}
\description{
Build a select matrix
}
