% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_select.R
\name{shift_matrix}
\alias{shift_matrix}
\title{shift_matrix}
\usage{
shift_matrix(compartments = NULL)
}
\arguments{
\item{compartments}{A vector of compartment}
}
\value{
A matrix
}
\description{
Build a shift matrix
}
