% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning.R
\name{SI.sum.pentype}
\alias{SI.sum.pentype}
\title{SI.sum.pentype}
\usage{
SI.sum.pentype(rows, Scomp, Icomp)
}
\arguments{
\item{rows}{A dataframe}

\item{Scomp}{A string}

\item{Icomp}{A string}
}
\value{
an integer, sum of animals
}
\description{
function to sum up animals from the 2 compartments combined in SI.pentype
}
