% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning.R
\name{sow.culling}
\alias{sow.culling}
\title{sow.culling}
\usage{
sow.culling(pen.id, culling_rate = 0.184)
}
\arguments{
\item{pen.id}{integer, vector of pen ids that are being weaned}

\item{culling_rate}{integer, rate of culling per weaning}
}
\value{
vector of integers (pen ids)
}
\description{
Function to calculate how many sows to remove at farrowing
 and which pens to take them from
}
\details{
According to Engblom et al. 2007, average annual removal rate for
sows was 49.5 % of these 26.9 % were of reproductive disorders.
If we assume, that the reproductive ones are removed at time of the gestation
(but the probability in the code hasn't been adjusted accordingly),
then 36.18 % of the sows are removed annually at the time of the weaning.
When the average amount of litters per sow is 2.23 (winpig 2019),
this gives the culling ratio of  ((36.18/2.23)/100 = 0.16 after each weaning
If one sows has average of 2.23 litters per year
(winpig 2019) According to Winpig 2019 statistics, the
average % of gilt litters in the herds was 24.2 %, this is fairly close to the
22.2 % culling that comes up when we get 2.23 litters per year
}
