% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{timers}
\alias{timers}
\title{timers}
\usage{
timers(
  pentype = c("Breeding", "Breeding dt", "Breeding buffer", "Gestation", "Gestation dt",
    "Farrowing", "Farrowing dt", "Growing", "Growing dt", "Growing buffer",
    "Growing buffer dt", "Finishing", "Finishing dt", "Gilt growing", "Gilt growing dt"),
  biweekly = FALSE
)
}
\arguments{
\item{pentype}{A string, dt = downtime}

\item{biweekly}{are the events biweekly. For the biweekly model we
have different downtimes between groups.}
}
\value{
an integer for setting countdown timer
}
\description{
Function to return appropriate countdown timer for the given
pentype
}
