% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning.R
\name{total.breeding}
\alias{total.breeding}
\title{total.breeding}
\usage{
total.breeding(farrowing.rate = 0.868, farrowing.pen.capacity = 22)
}
\arguments{
\item{farrowing.rate}{the proportion of sows that are bred that
successfully farrow}

\item{farrowing.pen.capacity}{the number of pens in a farrowing
room}
}
\value{
The number to breed in a week
}
\description{
Calculate the farrowing rate
the average farrowing rate for a sow/piglet according to Winpig
2019 is 86.8 %
}
