% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning.R
\name{weaning}
\alias{weaning}
\title{weaning}
\usage{
weaning(
  result,
  residual = NULL,
  events = NULL,
  fraction_to_gilt = 0.05,
  biweekly = FALSE
)
}
\arguments{
\item{result}{The result from the previous day or modified result
from the current day.}

\item{residual}{The rows with timer == 0 that still need to be
processed. The default is NULL if there is nothing to be done}

\item{events}{The events that have already been generated. The
default is NULL in the case that no events have yet been
generated for this day}

\item{fraction_to_gilt}{the proportion of pigs to be moved to gilt growing}

\item{biweekly}{Should the events we biweekly. Biweekly events
means that sows only farrow every second week and that at the
beginning of the model only 10 groups of gilts are added to
the model in 14 day increments.}
}
\value{
A list of the (modified) result, the new residual and the
    new events appended to the events you fed to the function
}
\description{
In this function we:
 - wean pigs to the grower and to the gilt grower.
 - wean sows to the breeding unit and cull sows that are to be culled
 - schedule events to bring in new gilts from gilt growing unit if we need refilling
 - refilling need is based on the total number of sows coming from farrowing to breeding and
 - sows in breeding buffer pens
All of these need to be together in one function as they have feedback on one one another.
}
